Flatten mocks/objects file structure
