--- server/src/test-utils/mocks/objects/legacy/clubs/index.ts
@@ -1,3 +1,0 @@
-export { mockClub } from './mock-club'
-export { mockClubDTO } from './mock-club-dto'
-export { mockClubEntity } from './mock-club-entity'
