--- server/src/test-utils/mocks/objects/legacy/events/index.ts
@@ -1,3 +1,0 @@
-export { mockEvent } from './mock-event'
-export { mockEventDTO } from './mock-event-dto'
-export { mockEventEntity } from './mock-event-entity'
