--- server/src/test-utils/mocks/objects/legacy/tags/index.ts
@@ -1,1 +1,0 @@
-export { mockTagEntity } from './mock-tag-entity'
