--- server/src/test-utils/mocks/objects/legacy/tags/mock-tag-entity.ts
@@ -1,9 +1,0 @@
-import { LegacyTagEntity } from '../../../../../shared/infra/db/entities/legacy/tag.entity'
-
-const mockTagEntity = (id: string): LegacyTagEntity => {
-  const tagEntity = new LegacyTagEntity()
-  tagEntity.name = `Tag Name ${id}`
-  return tagEntity
-}
-
-export { mockTagEntity }
