--- server/src/test-utils/mocks/objects/places/index.ts
@@ -1,3 +1,0 @@
-export { mockPlace } from './mock-place'
-export { mockPlaceDTO } from './mock-place-dto'
-export { mockPlaceEntity } from './mock-place-entity'
