--- server/src/test-utils/mocks/objects/users/index.ts
@@ -1,1 +1,0 @@
-export { mockUser } from './mock-user'
