--- server/src/test-utils/mocks/objects/washrooms/index.ts
@@ -1,3 +1,0 @@
-export { mockWashroom } from './mock-washroom'
-export { mockWashroomDTO } from './mock-washroom-dto'
-export { mockWashroomEntity } from './mock-washroom-entity'
